extracted strings
