+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".sharedpreferences.SplashActivity"
+            android:name=".sharedPreferences.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -23,7 +23,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".sharedpreferences.LoginActivity"
+            android:name=".sharedPreferences.LoginActivity"
             android:exported="false" />
         <activity
             android:name=".dataBinding.DataBindingLearning"
@@ -47,7 +47,7 @@
             android:name=".learningRecyclerviewAndAdapters.LearningRecyclerviewAndAdaptersActivity"
             android:exported="false" />
         <activity
-            android:name=".Assignment01_ConstraintLayoutPractice.Assignmnet01_UILayout"
+            android:name=".assignment01ConstraintLayoutPractice.Assignment01UILayout"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
