+++ app/src/main/java/com/example/myfirstapp/assignment01ConstraintLayoutPractice/Assignment01UILayout.kt
@@ -1,11 +1,11 @@
-package com.example.myfirstapp.Assignment01_ConstraintLayoutPractice
+package com.example.myfirstapp.assignment01ConstraintLayoutPractice
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import com.example.myfirstapp.R
 
-class Assignmnet01_UILayout : AppCompatActivity() {
+class Assignment01UILayout : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
