+++ app/src/main/res/layout/activity_assignmnet01_uilayout.xml
@@ -32,7 +32,6 @@
         android:hint="@string/email"
         android:inputType="textEmailAddress"
         android:paddingStart="10dp"
-        android:paddingLeft="10dp"
         android:paddingEnd="10dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -117,7 +116,7 @@
         android:layout_height="59dp"
         android:layout_marginTop="16dp"
         android:autofillHints=""
-        android:paddingLeft="10dp"
+        android:paddingStart="10dp"
         android:ems="10"
         android:textColorHint="@color/white"
         android:hint="@string/email"
@@ -177,7 +176,7 @@
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="8dp"
-        android:contentDescription="person image"
+        android:contentDescription="@string/person_image"
         android:scaleType="fitCenter"
         android:src="@mipmap/ic_launcher"
         app:layout_constraintEnd_toStartOf="@+id/textView"
