+++ app/src/main/res/layout/activity_login.xml
@@ -16,7 +16,7 @@
     android:textAlignment="gravity"
     android:textDirection="ltr"
     app:barrierMargin="23dp"
-    tools:context=".sharedpreferences.LoginActivity"
+    tools:context=".sharedPreferences.LoginActivity"
 
     >
 
@@ -30,7 +30,7 @@
 
         android:background="@drawable/button_background"
         android:clickable="false"
-        android:text="Login"
+        android:text="@string/login_button"
         android:textColor="@color/white"
         android:textColorLink="#FFFFFF"
         app:layout_constraintEnd_toEndOf="@+id/textInputLayout2"
@@ -65,7 +65,7 @@
             android:drawableStart="@android:drawable/ic_lock_lock"
             android:drawablePadding="12dp"
             android:drawableTint="#000000"
-            android:hint="Password"
+            android:hint="@string/password_string"
             android:textColor="@color/black"
 
 
@@ -108,7 +108,7 @@
             android:drawableTint="#000000"
             android:elegantTextHeight="false"
             android:fontFamily="sans-serif-medium"
-            android:hint="E-mail"
+            android:hint="@string/e_mail_string"
             android:inputType="textEmailAddress"
             android:singleLine="true"
             android:textColor="#000000"
@@ -122,7 +122,7 @@
         android:layout_marginStart="176dp"
         android:layout_marginTop="208dp"
         android:layout_marginEnd="177dp"
-        android:text="Login"
+        android:text="@string/login_button"
         android:textColor="#000000"
         android:textSize="48sp"
         android:textStyle="bold"
@@ -136,7 +136,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="8dp"
-        android:text="Welcome to login Screen"
+        android:text="@string/login_welcome"
         android:textSize="20sp"
         app:layout_constraintEnd_toEndOf="@+id/textView5"
         app:layout_constraintStart_toStartOf="@+id/textView5"
